+++ release-notes/release-notes.md
@@ -1,3 +1,38 @@
+## 🎉 1.4.72
+
+### ✨ Features added-
+
+1. Added `BubbleChart` component which receives data as an array of {x,y,r} values.
+Example-
+```js
+<BubbleChart
+   data={[
+      {x:20, y: 4, r: 10},
+      {x:40, y: 6, r: 20},
+      {x:30, y: 8, r: 60},
+      {x:120, y: 5, r: 40}
+   ]}
+/>
+```
+The props of Bubble chart are similar to those of Line chart. The `r` in the data array is analagous to `dataPointRadius` in the Line chart.
+See the props table here- https://github.com/Abhinandan-Kushwaha/react-native-gifted-charts/blob/master/docs/BubbleChart/BubbleChartProps.md
+
+### 🐛 Bug fixes
+
+1. Fixed the issue- "Combined Bar and Line chart, Line chart shifts upward on React Native if isAnimated is true". See https://github.com/Abhinandan-Kushwaha/react-native-gifted-charts/issues/1205
+
+2. Fixed the "Line chart coloring issue" for area charts, where gradient is unexpectedly applied to line. See https://github.com/Abhinandan-Kushwaha/react-native-gifted-charts/issues/1080
+
+3. FIxed the issue- "X-axis labels misalignment when using adjustToWidth and parentWidth in Stacked Bar Chart". See https://github.com/Abhinandan-Kushwaha/react-native-gifted-charts/issues/1203
+
+4. Fixed the issue- "In Line/Area charts verticalLineHeight not set correctly on data item level". See https://github.com/Abhinandan-Kushwaha/react-native-gifted-charts/issues/1202
+
+---
+
+---
+
+---
+
 ## 🎉 1.4.71
 
 ### ✨ Features added-